Replace usage of 'as' keyword
It hides that the conversion may be lossy.

This requires breaking ABI for the FFI interface, because the old
behaviour would truncate the number of masters if a plugin had more than
255 (which is not valid in-game, but esplugin could be given an invalid
plugin). This would ultimately cause memory management issues as the
caller wouldn't be able to pass back the full size of the array when it
came to freeing it.

The remaining usages were all to do with comparing with usize values
or indexing, and have been replaced by a function that will panic if the
u32 -> usize conversion will fail but which optimises away on platforms
where usize >= u32 (which is everywhere I expect this code to
run).

+++ ffi/src/plugin.rs
@@ -84,7 +84,7 @@ pub unsafe extern "C" fn esp_plugin_filename(
 pub unsafe extern "C" fn esp_plugin_masters(
     plugin_ptr: *const Plugin,
     plugin_masters: *mut *mut *mut c_char,
-    plugin_masters_size: *mut u8,
+    plugin_masters_size: *mut size_t,
 ) -> u32 {
     panic::catch_unwind(|| {
         if plugin_masters.is_null() || plugin_ptr.is_null() {
@@ -100,7 +100,7 @@ pub unsafe extern "C" fn esp_plugin_masters(
             c_string_vec.shrink_to_fit();
 
             *plugin_masters = c_string_vec.as_mut_ptr();
-            *plugin_masters_size = c_string_vec.len() as u8;
+            *plugin_masters_size = c_string_vec.len();
 
             mem::forget(c_string_vec);
